+++ animal-chipization/src/main/java/com/example/animalchipization/data/repositories/AnimalTypeRepository.java
@@ -1,7 +1,6 @@
 package com.example.animalchipization.data.repositories;
 
 import com.example.animalchipization.models.AnimalType;
-
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
