+++ animal-chipization/src/main/java/com/example/animalchipization/data/repositories/LocationPointRepository.java
@@ -1,7 +1,6 @@
 package com.example.animalchipization.data.repositories;
 
 import com.example.animalchipization.models.LocationPoint;
-
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
