+++ animal-chipization/src/main/java/com/example/animalchipization/exceptions/AccountWithThisEmailAlreadyExistsException.java
@@ -4,7 +4,4 @@ import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
 public class AccountWithThisEmailAlreadyExistsException extends RuntimeException {
-    public AccountWithThisEmailAlreadyExistsException(final String message) {
-        super(message);
-    }
 }
