+++ animal-chipization/src/main/java/com/example/animalchipization/models/AnimalVisitedLocation.java
@@ -1,11 +1,6 @@
 package com.example.animalchipization.models;
 
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
