+++ animal-chipization/src/main/java/com/example/animalchipization/services/annotations/CorrectGenderValidator.java
@@ -3,8 +3,8 @@ package com.example.animalchipization.services.annotations;
 import jakarta.annotation.Nullable;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
-import jakarta.validation.constraints.NotNull;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 
 @Service
@@ -16,7 +16,7 @@ public class CorrectGenderValidator implements ConstraintValidator<CorrectGender
 
     @Override
     public boolean isValid(@Nullable String gender, ConstraintValidatorContext constraintValidatorContext) {
-        if(gender != null) {
+        if (gender != null) {
             List<String> validValues = List.of("MALE", "FEMALE", "OTHER");
             return validValues.stream().anyMatch(gender::equals);
         } else {
