+++ animal-chipization/src/main/java/com/example/animalchipization/services/annotations/CorrectLifeStatus.java
@@ -2,18 +2,17 @@ package com.example.animalchipization.services.annotations;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
-import java.lang.annotation.Documented;
-import java.lang.annotation.Target;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
+
+import java.lang.annotation.*;
 
 @Documented
 @Constraint(validatedBy = CorrectLifeStatusValidator.class)
-@Target(value = { ElementType.PARAMETER })
+@Target(value = {ElementType.PARAMETER})
 @Retention(value = RetentionPolicy.RUNTIME)
 public @interface CorrectLifeStatus {
     String message() default "";
+
     Class<?>[] groups() default {};
+
     Class<? extends Payload>[] payload() default {};
 }
