+++ animal-chipization/src/main/java/com/example/animalchipization/services/annotations/CorrectLifeStatusValidator.java
@@ -4,6 +4,7 @@ import jakarta.annotation.Nullable;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 
 @Service
@@ -15,7 +16,7 @@ public class CorrectLifeStatusValidator implements ConstraintValidator<CorrectLi
 
     @Override
     public boolean isValid(@Nullable String lifeStatus, ConstraintValidatorContext constraintValidatorContext) {
-        if(lifeStatus != null) {
+        if (lifeStatus != null) {
             List<String> validValues = List.of("ALIVE", "DEAD");
             return validValues.stream().anyMatch(lifeStatus::equals);
         } else {
