+++ animal-chipization/src/main/java/com/example/animalchipization/web/controllers/AccountController.java
@@ -1,26 +1,21 @@
-package com.example.animalchipization.web;
+package com.example.animalchipization.web.controllers;
 
-import com.example.animalchipization.models.Account;
 import com.example.animalchipization.data.repositories.AccountRepository;
-
-import static com.example.animalchipization.data.specifications.AccountSpecification.*;
-
+import com.example.animalchipization.models.Account;
 import jakarta.validation.constraints.Min;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.NoSuchElementException;
 import java.util.Optional;
+
+import static com.example.animalchipization.data.specifications.AccountSpecification.*;
 
 @RestController
 @RequestMapping(path = "/accounts", produces = "application/json")
@@ -45,19 +40,10 @@ public class AccountController {
     }
 
     @GetMapping("/search")
-    public ResponseEntity<Iterable<Account>> searchForAccounts(
-            @RequestParam(name = "firstName", required = false) String firstName,
-            @RequestParam(name = "lastName", required = false) String lastName,
-            @RequestParam(name = "email", required = false) String email,
-            @RequestParam(name = "from", required = false, defaultValue = "0") @Min(0) Integer from,
-            @RequestParam(name = "size", required = false, defaultValue = "10") @Min(1) Integer size) {
+    public ResponseEntity<Iterable<Account>> searchForAccounts(@RequestParam(name = "firstName", required = false) String firstName, @RequestParam(name = "lastName", required = false) String lastName, @RequestParam(name = "email", required = false) String email, @RequestParam(name = "from", required = false, defaultValue = "0") @Min(0) Integer from, @RequestParam(name = "size", required = false, defaultValue = "10") @Min(1) Integer size) {
 
         PageRequest pageRequest = PageRequest.of(from, size, Sort.by("id").ascending());
-        Specification<Account> specifications = Specification.where(
-                firstNameLike(firstName)
-                        .and(lastNameLike(lastName))
-                        .and(emailLike(email))
-        );
+        Specification<Account> specifications = Specification.where(firstNameLike(firstName).and(lastNameLike(lastName)).and(emailLike(email)));
         Iterable<Account> accounts = accountRepository.findAll(specifications, pageRequest).getContent();
         return new ResponseEntity<>(accounts, HttpStatus.valueOf(200));
     }
