+++ animal-chipization/src/main/java/com/example/animalchipization/web/controllers/AnimalController.java
@@ -1,10 +1,7 @@
-package com.example.animalchipization.web;
+package com.example.animalchipization.web.controllers;
 
-import com.example.animalchipization.models.Animal;
 import com.example.animalchipization.data.repositories.AnimalRepository;
-
-import static com.example.animalchipization.data.specifications.AnimalSpecification.*;
-
+import com.example.animalchipization.models.Animal;
 import com.example.animalchipization.services.annotations.CorrectGender;
 import com.example.animalchipization.services.annotations.CorrectLifeStatus;
 import jakarta.validation.constraints.Min;
@@ -15,15 +12,13 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.util.NoSuchElementException;
 import java.util.Optional;
+
+import static com.example.animalchipization.data.specifications.AnimalSpecification.*;
 
 @RestController
 @RequestMapping(path = "/animals", produces = "application/json")
